fix: gpt api 분리 및 코드 수정
