+++ backend/asgi.py
@@ -10,13 +10,13 @@ https://docs.djangoproject.com/en/5.0/howto/deployment/asgi/
 import os
 from channels.routing import ProtocolTypeRouter, URLRouter
 from django.core.asgi import get_asgi_application
-import chat.routing
+import gpt.routing
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "backend.settings")
 
 application = ProtocolTypeRouter({
     "http": get_asgi_application(),
     "websocket": URLRouter(
-        chat.routing.websocket_urlpatterns
+        gpt.routing.websocket_urlpatterns
     ),
 })
