+++ chat/serializers.py
@@ -1,11 +1,6 @@
 from rest_framework import serializers
 
 
-class GetGPTTalkSerializer(serializers.Serializer):
-    character_id = serializers.IntegerField()
-    episode_id = serializers.IntegerField()
-
-
-class GetGPTAnswerSerializer(serializers.Serializer):
+class AnswerSerializer(serializers.Serializer):
     choice_content = serializers.CharField(max_length=1000)
     mz_percent = serializers.IntegerField()
