+++ chat/urls.py
@@ -5,6 +5,5 @@ from . import views
 appname = "chat"
 
 urlpatterns = [
-    path('gpt-talk', views.GetGPTTalkView.as_view(), name='gpt-talk'),
-    path('gpt-answer', views.GetGPTAnswerView.as_view(), name='gpt-answer'),
+    path('answer', views.AnswerView.as_view(), name='choice'),
 ]
