+++ chat/views.py
@@ -1,63 +1,39 @@
-import redis
+import requests
 from django.http import JsonResponse
 from drf_yasg.utils import swagger_auto_schema
 from rest_framework import status
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.reverse import reverse
 from rest_framework.views import APIView
-from rest_framework_simplejwt.authentication import JWTAuthentication
 
-from .serializers import GetGPTTalkSerializer, GetGPTAnswerSerializer
-from .tasks import get_gpt_talk, get_gpt_answer
-
-talk_count = 1
+from .serializers import AnswerSerializer
 
 
-class GetGPTTalkView(APIView):
+class AnswerView(APIView):
     # authentication_classes = [JWTAuthentication]
     # permission_classes = [IsAuthenticated]
+
     @swagger_auto_schema(
-        request_body=GetGPTTalkSerializer,
-        operation_id='GPT의 대사를 가져오는 API',
+        operation_id='답변과 피드백을 가져오는 API',
+        request_body=AnswerSerializer,
     )
     def post(self, request):
-        global talk_count
-        serializer = GetGPTTalkSerializer(data=request.data)
-        if serializer.is_valid():
-            character_id = serializer.validated_data.get('character_id')
-            episode_id = serializer.validated_data.get('episode_id')
-            redis_key = "count"
-
-            r = redis.Redis(host='redis', port=6379, db=0)
-            if r.exists('episode_id'):
-                r.delete('episode_id')
-            r.set('episode_id', episode_id)
-            r.set(redis_key, talk_count)
-            r.set('character_id', character_id)
-            talk_count += 1
-            result = get_gpt_talk.delay(character_id, episode_id)
-            return JsonResponse({'task_id': result.task_id}, status=status.HTTP_202_ACCEPTED)
-        else:
-            return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-class GetGPTAnswerView(APIView):
-    # authentication_classes = [JWTAuthentication]
-    # permission_classes = [IsAuthenticated]
-    @swagger_auto_schema(
-        request_body=GetGPTAnswerSerializer,
-        operation_id='GPT의 대답을 가져오는 API',
-    )
-    def post(self, request):
-        serializer = GetGPTAnswerSerializer(data=request.data)
-        r = redis.Redis(host='redis', port=6379, db=0)
+        serializer = AnswerSerializer(data=request.data)
         if serializer.is_valid():
             choice_content = serializer.validated_data.get('choice_content')
             mz_percent = serializer.validated_data.get('mz_percent')
-            character_id = r.get('character_id')
-            episode_id = r.get('episode_id')
-            talk_content = r.get('talk_content').decode('utf-8')
-
-            result = get_gpt_answer.delay(choice_content, character_id, episode_id, talk_content, mz_percent)
-            return JsonResponse({'task_id': result.task_id}, status=status.HTTP_202_ACCEPTED)
+            url = request.build_absolute_uri(reverse('gpt:gpt-answer'))
+            payload = {
+                "choice_content": choice_content,
+                "mz_percent": mz_percent
+            }
+            headers = {
+                'Content-Type': 'application/json',
+                # 'Authorization': 'Bearer ' + request.headers['Authorization']
+            }
+            response = requests.post(url, json=payload, headers=headers)
+            if response.status_code == status.HTTP_202_ACCEPTED:
+                return JsonResponse({'task_id': response.json().get('task_id')}, status=status.HTTP_200_OK)
+            else:
+                return JsonResponse(response.json(), status=response.status_code)
         else:
             return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
