+++ gpt/routing.py
@@ -1,5 +1,5 @@
 from django.urls import re_path
-from chat import consumers
+from gpt import consumers
 
 websocket_urlpatterns = [
     re_path(r'ws/gpt/$', consumers.ChatConsumer.as_asgi()),
